added permission for all users

+++ backend/src/resolvers/Mutation.js
@@ -3,6 +3,8 @@ const jwt = require("jsonwebtoken");
 const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
+
 const Mutation = {
   async createItem(
     parent,
@@ -62,25 +64,28 @@ const Mutation = {
     );
     return item;
   },
-
   async deleteItem(
     parent,
     { id },
     {
+      req,
       db: { query, mutation },
     },
     info
   ) {
     // 1. find the item
-    const item = await query.item(
-      { where: { id } },
-      `{
-      id
-      title
-    }`
-    );
+    const item = await query.item({ where: { id } }, `{ id title user { id }}`);
+
     // 2.check they own that item, or have the permisions
-    // TODO:
+    const ownsItem = item.user.id === req.userId;
+    const hasPermissions = req.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
+
     // 3. Delete it
     return mutation.deleteItem({ where: { id } }, info);
   },
@@ -231,6 +236,40 @@ const Mutation = {
 
     // 8. return the new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.req.userId) {
+      throw new Error("You must be logged in!");
+    }
+
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.req.userId,
+        },
+      },
+      info
+    );
+
+    // 3. Check if they have permissions to do this
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: {
+          id: args.userId,
+        },
+      },
+      info
+    );
   },
 };
